koji: fix build tag and add brew id to oscontainer
It looks like the Koji Content generator never did anything with the tag
that was passed in from the pipeline.  Hopefully, all it takes to tag a
build is to use the KOJI_TAG environment variable.

Now that we have the koji/brew build ID available, lets add it to the
oscontainer so we can reference brew builds on the OpenShift release
browser.