tests: choose latest y-stream versioning for rhcos upgrade
The way this test determined the latest released OCP version was
finding a node without an upgrade edge in the cincinnati graph.
There was an assumption that only one such node existed, however,
two such nodes can exist when an upgrade path is blocked if it is
discovered that the upgrade causes breakage. If two nodes exist
without an upgrade edge, chooser the newer one.